# openai-whisper-cpp: use impure native optimizations