@FliegendeWurst 2024-11-22
Here is what you can try: build the package with optimizations enabled (by some other means, maybe a configure flag) and check whether it can also run on older CPUs (try a VM). Some packages still add runtime checks, even if you enable avx etc.